Updates to zend-expressive-router 2.4.0
Only change in that release is that the `Route` instance now triggers a
deprecation notice if non-MiddlewareInterface `$middleware` is provided
at instantiation. This patch updates the tests to mock the
`MiddlewareInterface` and provide it as an instance whenever a new
`Route` instance is created. References to the `'middleware'` default in
route specifications also reference it.